Inline messaging-runtime into panel; extract pure utils
Fold the one-call messaging-runtime wrapper into ConversationPanel as a
direct createD1MessageRepositoryFromEnv() and delete the file.

Extract stateless timestamp/attachment helpers out of the bloated panel
into utils/format-timestamp.ts and utils/attachments.ts.

Mark unused conversation.actions exports with TODOs.
