+++ src/features/messaging-next/conversation.actions.ts
@@ -109,11 +109,13 @@ export function createConversationActions(store: ConversationStateStore) {
     setState('messages', (m) => m.id === tempId, 'status', 'failed');
   }
 
+  // TODO: integrate or remove
   function updateReactions(messageId: string, reactions: ReactionSummary[]) {
     const idx = state.messages.findIndex((m) => m.id === messageId);
     if (idx !== -1) setState('messages', idx, 'reactions', reactions);
   }
 
+  // TODO: integrate or remove
   function addSystemMessage(text: string, actions?: MessageAction[]) {
     const id = crypto.randomUUID();
     setState('messages', (msgs) => [
@@ -134,6 +136,7 @@ export function createConversationActions(store: ConversationStateStore) {
     return id;
   }
 
+  // TODO: integrate or remove
   function removeSystemMessage(id: string) {
     setState('messages', (msgs) => msgs.filter((m) => m.id !== id));
   }
@@ -146,6 +149,7 @@ export function createConversationActions(store: ConversationStateStore) {
     setState('isPendingPrivateResponse', v);
   }
 
+  // TODO: integrate or remove
   function markAllRead() {
     setState('unreadCount', 0);
   }
