--- src/features/messaging-next/messaging-runtime.ts
@@ -1,9 +1,0 @@
-import { createD1MessageRepositoryFromEnv } from '../../stores/message-repository';
-
-// Messages persist + live-push through the data worker (D1 + ConversationChannel
-// DO). Conversation registry/membership is resolved through the data worker.
-export function createMessagingRuntime() {
-  return {
-    messageRepository: createD1MessageRepositoryFromEnv(),
-  };
-}
